pyproject.toml
[project]
name = "llmstxt-standalone"
version = "0.1.1"
description = "Generate llms.txt from built HTML documentation"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["llms", "documentation", "markdown", "mkdocs"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Documentation",
    "Typing :: Typed",
]

authors = [{ name = "Shaan Majid", email = "shaanmajid64@gmail.com" }]

dependencies = [
    "typer>=0.9.0",
    "pyyaml>=6.0",
    "ruamel.yaml>=0.18",
    "beautifulsoup4>=4.12",
    "markdownify>=0.14,<2.0",
    "mdformat>=0.7,<2.0",
    "mdformat-tables>=1.0",
]

[project.scripts]
llmstxt-standalone = "llmstxt_standalone.cli:app"

[project.urls]
Repository = "https://github.com/shaanmajid/llmstxt-standalone"
Issues = "https://github.com/shaanmajid/llmstxt-standalone/issues"

[dependency-groups]
dev = [
    { include-group = "hooks" },
    { include-group = "lint" },
    { include-group = "typecheck" },
    { include-group = "test" },
]
hooks = ["prek>=0.3.0"]
lint = ["ruff>=0.14.14"]
typecheck = ["ty>=0.0.14"]
test = ["pytest>=8.0.0", "pytest-cov>=6.0.0"]

[tool.uv]
# renovate: datasource=pypi depName=uv
required-version = ">=0.9.27"

[build-system]
# renovate: datasource=pypi depName=uv_build
requires = ["uv_build>=0.9.0,<=0.9.27"]
build-backend = "uv_build"
